requirements.txt
streamlit
torch
torchvision
timm
pillow
pandas

